% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luar-package.R
\docType{package}
\name{luar-package}
\alias{luar}
\alias{luar-package}
\title{luar: Experimental embedding of lua in R}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noamross/luar}
  \item Report bugs at \url{https://github.com/noamross/luar/issues}
}

}
\author{
\strong{Maintainer}: Noam Ross \email{noam.ross@gmail.com}

}
\keyword{internal}
